backend/requirements.txt
fastapi
uvicorn
python-docx
pymupdf
python-multipart